feat: Split out exercise form and fix datepicker
