+++ src/app/components/ExercisesList/ExercisesList.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from "react";
 import { IExercise } from "@/app/api/user/exercises/route";
-import LogExerciseForm from "@/app/components/LogExerciseForm";
+import LogExerciseForm from "../LogExerciseModal";
 import useUserExercises from "@/app/hooks/useUserExercises";
 import Link from "next/link";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
