+++ src/app/components/Forms/index.ts
@@ -1,3 +1,4 @@
 import SearchForUserFormLarge from "./SearchForUserFormLarge/SearchForUserFormLarge";
+import LogExerciseForm from "./LogExerciseForm/LogExerciseForm";
 
-export { SearchForUserFormLarge };
+export { SearchForUserFormLarge, LogExerciseForm };
