--- src/app/components/LogExerciseForm/index.ts
@@ -1,3 +1,0 @@
-import LogExerciseForm from "./LogExerciseForm";
-
-export default LogExerciseForm;
