+++ src/app/components/LogExerciseModal/index.ts
@@ -1,0 +1,3 @@
+import LogExerciseModal from "./LogExerciseModal";
+
+export default LogExerciseModal;
